+++ src/App.js
@@ -10,7 +10,7 @@ function App() {
       <Routes>
         <Route path = "/" element={<Home />}></Route>
         <Route path = "/login" element={<Login />}></Route>
-        <Route path="/oauth" element={<OAuth />} /> {/* 여기 경로 수정하면 백한테 주소 리다이랙트 주소 수정 요청필요 */}
+        <Route path="/oauth" element={<div>Loading...</div>}/*{<OAuth />}*/ /> {/* 여기 경로 수정하면 백한테 주소 리다이랙트 주소 수정 요청필요 */}
       </Routes>
     </BrowserRouter>
   );
